Make cmake aware of existing open-source builds.
I have to wonder if anyone else has suffered through building openssl
too many times. At least on windows platform, the library detection
does not work, so even when openssl and curl are built, cmake rebuilds
the darn thing over and over.

Now, openssl is especially weird here because it should be building
libssld.dll but it build libssl.dll, during a DEBUG configure in cmake.
Not sure why. But it causes the library detection to fail. So, I'm
manually adding the release name, libssl.

Curl is simply a case of the wrong libname being looked for, I see
libcurl-d_imp.dll and by default cmake's libcurl finder is only looking
for curl and maybe libcurl.