Libcurl was also building in debug mode.
I don't know if this is the right choice. But in order to distribute
these libs they have to be in release mode. Otherwise they have
dependencies on things like vcruntime140D.dll, which are only present
on the building dev machine, and are not considered portable.

So, I made the 3rd party deps build as release. But, maybe that's not
the right thing to do? Are these dependent libs built with the producer
supposed to be ditrubted with it or are they just for development? The
writers of the build system in the aws producers seem ambivalent.
This...

Works.